Fix for libstdc++ building with mshort & mfastcall
mshort & mfastcall are not available in mintlib.
Do not try to configure libstdc for it,
this will fail eg. when m5475/mshort is configured,
and that library is not found. The link test will then
use the libc.a for m68000, which contains references
to __ltxf2, which is not compiled for coldfire