Make mintelf toolchain use emutls
The elf assembler will successfully accept the code fragment used for
the test, but that will later cause the compiler to emit library calls
to __m68k_read_tp, which does not exist for our target in libgcc.a. As
a result, checking for TLS support for the C++ compiler will fail, and
the thread_local keyword not being supported.